+++ src/test/kotlin/com/zepben/evolve/services/network/tracing/connectivity/LimitedConnectedEquipmentTraceTest.kt
@@ -11,6 +11,7 @@ package com.zepben.evolve.services.network.tracing.connectivity
 import com.zepben.evolve.cim.iec61970.base.core.ConductingEquipment
 import com.zepben.evolve.cim.iec61970.base.core.Terminal
 import com.zepben.evolve.cim.iec61970.base.wires.Junction
+import com.zepben.evolve.services.network.tracing.Tracing.normalConnectedEquipmentTrace
 import com.zepben.evolve.services.network.tracing.feeder.FeederDirection
 import com.zepben.evolve.testing.TestNetworkBuilder
 import com.zepben.testutils.junit.SystemLogExtension
@@ -141,7 +142,7 @@ internal class LimitedConnectedEquipmentTraceTest {
         trace.run(listOf(simpleNs["b2"]!!), 2, FeederDirection.DOWNSTREAM)
 
         verify(traversal).run(any<ConductingEquipment>(), any())
-        verify(traversal).run(simpleNs["c3"]!!, false)
+        verify(traversal).run(simpleNs["c3"]!!)
     }
 
     @Test
@@ -149,7 +150,7 @@ internal class LimitedConnectedEquipmentTraceTest {
         trace.run(listOf(simpleNs["b2"]!!), 2, FeederDirection.UPSTREAM)
 
         verify(traversal).run(any<ConductingEquipment>(), any())
-        verify(traversal).run(simpleNs["c1"]!!, false)
+        verify(traversal).run(simpleNs["c1"]!!)
     }
 
     @Test
@@ -170,8 +171,8 @@ internal class LimitedConnectedEquipmentTraceTest {
         trace.run(listOf(ns["j2"]!!), 2, FeederDirection.BOTH)
 
         verify(traversal, times(2)).run(any<ConductingEquipment>(), any())
-        verify(traversal).run(ns["c1"]!!, false)
-        verify(traversal).run(ns["c5"]!!, false)
+        verify(traversal).run(ns["c1"]!!)
+        verify(traversal).run(ns["c5"]!!)
     }
 
     @Test
@@ -193,7 +194,7 @@ internal class LimitedConnectedEquipmentTraceTest {
         trace.run(listOf(ns["j2"]!!), 2, FeederDirection.NONE)
 
         verify(traversal).run(any<ConductingEquipment>(), any())
-        verify(traversal).run(ns["c5"]!!, false)
+        verify(traversal).run(ns["c5"]!!)
     }
 
     @Test
@@ -246,6 +247,23 @@ internal class LimitedConnectedEquipmentTraceTest {
 
         assertThat(results, aMapWithSize(1))
         assertThat(results[ns["j0"]], equalTo(0))
+    }
+
+    @Test
+    internal fun withDirectionCanStopOnStartItem() {
+        val ns = TestNetworkBuilder()
+            .fromJunction(numTerminals = 1) // j0
+            .toAcls() // c1
+            .toJunction(numTerminals = 1) // j2
+            .addFeeder("j0")
+            .build()
+
+        val lcet = LimitedConnectedEquipmentTrace({ normalConnectedEquipmentTrace() }) { it.normalFeederDirection }
+        val results = lcet.run(listOf(ns["j0"]!!), 1, FeederDirection.DOWNSTREAM)
+
+        assertThat(results, aMapWithSize(2))
+        assertThat(results[ns["j0"]], equalTo(0))
+        assertThat(results[ns["c1"]], equalTo(1))
     }
 
     @Test